Extract runtime authority panel
